- 조회 시 즉시 DB를 업데이트하지 않고, Redis의 INCR 명령어를 사용하여 실시간 조회수를 임시 저장
- INCR 사용 시 발생하는 직렬화 오류를 방지하고자 StringRedisTemplate을 사용하여 수정
- 5분 간격(Cron 설정)으로 동작하는 스케줄러를 구현하여 Redis에 쌓인 조회수를 DB에 반영
- 카운트 컬럼을 int 타입을 long으로 변경
- 테스트 코드 추가

+++ src/main/java/com/project/market_service/product/domain/Product.java
@@ -62,10 +62,10 @@ public class Product extends BaseEntity {
     private ProductStatus status;
 
     @Column(nullable = false)
-    private int viewCount = 0;
+    private long viewCount = 0;
 
     @Column(nullable = false)
-    private int wishCount = 0;
+    private long wishCount = 0;
 
     @Column(nullable = false, columnDefinition = "POINT SRID 4326")
     private Point location;
@@ -122,8 +122,8 @@ public class Product extends BaseEntity {
         this.description = description;
         this.price = price;
         this.status = status;
-        this.viewCount = 0;
-        this.wishCount = 0;
+        this.viewCount = 0L;
+        this.wishCount = 0L;
         this.location = location;
         this.address = address;
     }
